Split the comparative view for better overview
